delete and add buttons using jquery
