+++ .jshintrc
@@ -1,0 +1,11 @@
+{
+  "predef": [ "document", "$", "console", "angular", "jQuery", "require" ],
+  "esversion": 6,
+  "globalstrict": true,
+  "curly": true,
+  "eqeqeq": true,
+  "undef": true,
+  "unused": true,
+  "browserify": true
+  }
+}
