+++ index.html
@@ -4,8 +4,8 @@
 <head>
   <meta charset="utf-8">
   <title>Music History</title>
-  <link rel="stylesheet" type="text/css" href="styles.css">
-  <link rel="stylesheet" type="text/css" href="musicHistory4.css">
+  <link rel="stylesheet" type="text/css" href="css/styles.css">
+  <link rel="stylesheet" type="text/css" href="css/musicHistory5.css">
 </head>
 <body>
 
@@ -80,7 +80,6 @@
 
 		<!-- right section -->
 
-
 		<div class="right">
 		<div id="inputSongs"></div>
 		<div id="inputSongs2"></div>
@@ -90,6 +89,7 @@
  		</div>
  		
 	</article> 
-	<script src="songs4Part2.js"></script>
+  	<script src="node_modules/jquery/dist/jquery.min.js"></script>
+		<script src="src/main.js"></script>
 </body>
 </html>
